Improve auto-reload and fix concurrency issues
Auto-reloading now uses a timer that automatically checks for changes to the file once per second. The previous implementation used FileSystemWatcher which is a nasty bastard of a class that loves to randomly not work for no reason, among other problems.

In a related change, we now use a lock when reading or writing data to disk, to ensure we never get sharing violations.