Update TargetFramework of JECS
Using 'netstandard2.0' can drag some projects down. And there are a couple of benefits of using 'netstandard2.1'. The recent Unity versions also support netstandard2.1.

Removing 'System.ValueTuple' dependency, cause that is included in the new target framework.
Also removing 'Microsoft.CSharp' - I actually do not know what it was used for. But it does not seem to be required for compilation. Removed it.