Ensure top-level key-nodes have no indentation
Previously JECS would allow top-level key lines to have indentation. Which allows top-level key nodes with higher indentation than following top-level key-nodes - which makes no sense.
According to the Wiki of JECS, top-level key-nodes should have an indentation of zero.

This commit adds a check for this, ensuring that the indentation is zero.
It also adds a test to prevent this from happening in future.