Refactor JECS tree parser
The DataConverter class had a big method dedicated to string parsing.
That has been split up into multiple methods and in general refactored to be less confusing (I hope). I also hope that now you do no longer need gods guidance.

The core changes are the usage of an Enumerator instead of a loop over lines. This allows the multi-line string logic to grab new lines.
And the addition of a 'ParserData' struct, which is passed around to limit parameter count. But more importantly, it holds the line enumerator and the stack and provides helpers for both. This reduces the clutter in the rest of the code.

One could argue to make the whole parsing methods part of that struct - which would work. But it does not matter too much. I think the current solution is fairly clean. Could surely be better, but works.