Add new API methods & rework API implementation of DataFile
The DataFile classes as well as DistributedData all use the same set of methods to allow access to underlying nodes. However some methods had been missing, such as:
- ReadableDataFile: TryGetAtPath / TryGetAtPathNonGeneric
- ReadableWritableDataFile: ResetValueAtPathToDefault
- DistributedData: GetAtPath(key) for easier overload usage / TryGetAtPath / TryGetAtPathNonGeneric

Besides that the implementation of each method in these classes have been revisited and improved (subjectively). For DataFile 'TryGetNode' & 'TryGetNodeAtPath' have been added, which many other methods can use.
The optimizations tried to reduce looping over the same data, by utilizing Try methods at their core. Another main reason of switching to Try is that each method can deal with it's default values as seen fit, and thus reducing the need to pass down the default values on the call stack => More flexibility.

Besides that this commit fixes:
- 'ResetValueToDefault' breaking when no default was set.
- 'DeleteKeyAtPath' creating nodes, while it was supposed to delete a nested child within the missing nodes.

For 'DeleteKeyAtPath' a new test was created. And for the very complicated and new 'ResetValueAtPathToDefault' a test was added as well (and it was very necessary :P).