Fix GetAsObject() overwriting default class field values
By JECS API rules, when no default is provided a language default is to be used.
For objects/classes an instance is created and returned. Only fields which are present in JECS should be overwritten.

The previous code however overwrote all fields, without an appropriate default. For nullables it would assign 'null', ignoring if the instance by default sets a different value.
The fix is to simply use a newly added API method and only overwrite the value when a JECS value exists.

This commit also changes the other GetAsObject method with default to use a Try method. This has no functional change.