Fix ReadableWritableDataFile not setting unset values
The getters in ReadableWritableDataFile relied uppon the parent class ReadableDataFile delegating the method calls to the NonGeneric getter with default values.
Only these two methods (with & without path) got overriden to automatically set values which are unset.

My last PR overlooked this behavior and changed that every method delegates to these two overridden methods.

This commit fixes this again, by not relying on the parent class implementation detail, but instead implementing an override for every getter - which ensures that unset values are set.
This has the nice benefit, that now the XML documentation exists for every getter (now mentioning the unset set behavior).

Adjusted the XML documentation of all the important getters and setters to be uniform.