Fix GetAsObject not behaving consistently with the rest of the API
JECS has two concepts:
- No default value provided: API tries to find: 1) In data. 2) In provided default JECS text. 3) Generates a language default.
- Default value provided: API tries to find in data, else it returns the provided default.

The current GetAsObject, was recently split. To actually support both API method types. The implementation for no-provided-default however did not search in the provided default JECS text when resolving each field (it did in the other cases).

This commit addresses this and makes it search in the default JECS text.