Add WrappedObjectDataFile
This file is moved from a project using JECS into JECS. It was originally written by the JECS author Jimmy.

There are many applications where a single object is stored as a single JECS file. This new class makes it easy to load this data and work with it - reducing code overhead.

Due to it's structure 'Save()' has to be called explicitly to sync the data back into the JECS file. But unless working with JECS explicitly this has to be done anyway.