Fix oversight issue caused by 'readonly' keyword
Only trust an IDE as far as you can uninstall it. My IDE recommended to add 'readonly' which makes ton of sense for objects addressed by reference. But generics can be structs as well - and a struct is modified by copy and updating the reference - hence it may not be readonly.

Point being, I broke the file I just added - by trying to clean up something. This commit undoes that.